+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - The so called "perfect eq" preset. Not that there really is something we can
 call perfect equalizer preset. But people like to use it as a starting point.
 So I added it.
+- Presets menu label shows the last selected preset when PulseEffects is
+restarted.
 
 ### Removed
 - Expander plugin. Reason: I thought it was open source but it is proprietary.